Switch to use convention for loading data bag items
This switches the chef-support method to use the metadata about the
current example to determine where to load data bag json items from.

There are some limitations to this approach, so it can be improved, but
this greatly simplifies setup until you need to override the convention.
This may be a better starting point for this library.

[26059]