Add support for chefspec 4.0
In order to support chefspec 4.0 and rspec 3.0, this commit updates the
dependencies to now require chefspec 4.0. In addition, due to a change
in rspec, the way we access the current spec metadata needed to be
changed to account for the example method no longer being available
within a spec. Instead we use RSpec.current_example as this is set
before each spec to the correct metadata for that spec.

[12800973138385](https://app.asana.com/0/0/12800973138385)