Remove bundler from gemspec and add gitignore file
Since it is not necessary to lock bundler in the gemspec, this
pull request removes the version lock specified in the gemspec for
bundler, as well as the bundler requirement in the spec_helper.

This pull request also removes the Gemfile.lock and adds a gitignore
file in order to prevent commits of future Gemfile.lock files.

[1120378415217082](https://app.asana.com/0/0/1120378415217082/f)