Setup github action workflow for github gem repo
In order to use the github package repo for this gem as well as the
github action for pushing the gem, this pull request adds the
`gem-push.yml` file.

This pull request also adds a gem-version-check file to enforce the
bumping of gem versions on branch builds.

This pull request also includes a bump to the gem version.